Move encoding tests
