+++ packages/backend/.mocharc.json
@@ -3,5 +3,6 @@
   "require": ["esbuild-register"],
   "watchExtensions": "ts",
   "extension": "ts",
-  "file": "src/test/setup.ts"
+  "file": "src/test/setup.ts",
+  "reporterOption": "maxDiffSize=0"
 }
