+++ packages/crypto/.mocharc.json
@@ -2,5 +2,6 @@
   "spec": "{src,test}/**/*.test.ts",
   "require": ["esbuild-register"],
   "watchExtensions": "ts",
-  "extension": "ts"
+  "extension": "ts",
+  "reporterOption": "maxDiffSize=0"
 }
