+++ packages/encoding/package.json
@@ -2,8 +2,8 @@
   "name": "@explorer/encoding",
   "version": "0.1.0",
   "private": true,
-  "main": "build/src/index.js",
-  "types": "build/src/index.d.ts",
+  "main": "build/index.js",
+  "types": "build/index.d.ts",
   "sideEffects": false,
   "engines": {
     "node": "^16.0.0",
@@ -12,12 +12,13 @@
   "scripts": {
     "format": "prettier --check .",
     "format:fix": "prettier --write .",
-    "lint": "eslint --ext .ts --max-warnings 0 src test",
+    "lint": "eslint --ext .ts --max-warnings 0 src",
     "lint:fix": "yarn lint --fix",
     "typecheck": "tsc --noEmit",
+    "prebuild": "yarn clean",
     "build": "tsc",
-    "test": "mocha",
-    "get-onchain-data": "node -r esbuild-register src/data"
+    "clean": "rm -rf build",
+    "test": "mocha"
   },
   "dependencies": {
     "@explorer/types": "*"
