+++ packages/encoding/src/decoding/readConditions.test.ts
@@ -1,10 +1,10 @@
 import { PedersenHash } from '@explorer/types'
 import { expect } from 'earljs'
 
-import { DecodingError } from '../../src'
-import { readConditions } from '../../src/decoding/readConditions'
-import { ByteWriter } from '../../src/encoding/ByteWriter'
-import { readToDecode } from './readToDecode'
+import { ByteWriter } from '../encoding/ByteWriter'
+import { readToDecode } from '../test/readToDecode'
+import { DecodingError } from './DecodingError'
+import { readConditions } from './readConditions'
 
 describe(readConditions.name, () => {
   const decode = readToDecode(readConditions)
