+++ packages/encoding/src/decoding/readOraclePrices.test.ts
@@ -1,11 +1,11 @@
 import { AssetId } from '@explorer/types'
 import { expect } from 'earljs'
 
-import { DecodingError } from '../../src'
-import { readOraclePrices } from '../../src/decoding/readOraclePrices'
-import { ByteWriter } from '../../src/encoding/ByteWriter'
-import { encodeAssetId } from '../../src/encoding/encodeAssetId'
-import { readToDecode } from './readToDecode'
+import { ByteWriter } from '../encoding/ByteWriter'
+import { encodeAssetId } from '../encoding/encodeAssetId'
+import { readToDecode } from '../test/readToDecode'
+import { DecodingError } from './DecodingError'
+import { readOraclePrices } from './readOraclePrices'
 
 describe('readOraclePrices', () => {
   const decode = readToDecode(readOraclePrices)
