+++ packages/encoding/src/encoding/encodeOnChainData.test.ts
@@ -7,27 +7,29 @@ import {
 } from '@explorer/types'
 import { expect } from 'earljs'
 
-import { decodeOnChainData, encodeOnChainData, OnChainData } from '../../src'
-import { ByteReader } from '../../src/decoding/ByteReader'
-import { readAssetConfigHashes } from '../../src/decoding/readAssetConfigHashes'
-import { readConditions } from '../../src/decoding/readConditions'
-import { readForcedActions } from '../../src/decoding/readForcedActions'
-import { readFundingEntries } from '../../src/decoding/readFundingEntries'
-import { readFundingIndices } from '../../src/decoding/readFundingIndices'
-import { readModifications } from '../../src/decoding/readModifications'
-import { readOraclePrices } from '../../src/decoding/readOraclePrices'
-import { readPositionUpdate } from '../../src/decoding/readPositionUpdate'
-import { readState } from '../../src/decoding/readState'
-import { ByteWriter } from '../../src/encoding/ByteWriter'
-import { encodeForcedActions } from '../../src/encoding/encodeForcedActions'
-import { writeAssetConfigHashes } from '../../src/encoding/writeAssetConfigHashes'
-import { writeConditions } from '../../src/encoding/writeConditions'
-import { writeFundingEntries } from '../../src/encoding/writeFundingEntries'
-import { writeFundingIndices } from '../../src/encoding/writeFundingIndices'
-import { writeModifications } from '../../src/encoding/writeModifications'
-import { writeOraclePrices } from '../../src/encoding/writeOraclePrices'
-import { writePositionUpdate } from '../../src/encoding/writePositionUpdate'
-import { writeState } from '../../src/encoding/writeState'
+import { ByteReader } from '../decoding/ByteReader'
+import { decodeOnChainData } from '../decoding/decodeOnChainData'
+import { readAssetConfigHashes } from '../decoding/readAssetConfigHashes'
+import { readConditions } from '../decoding/readConditions'
+import { readForcedActions } from '../decoding/readForcedActions'
+import { readFundingEntries } from '../decoding/readFundingEntries'
+import { readFundingIndices } from '../decoding/readFundingIndices'
+import { readModifications } from '../decoding/readModifications'
+import { readOraclePrices } from '../decoding/readOraclePrices'
+import { readPositionUpdate } from '../decoding/readPositionUpdate'
+import { readState } from '../decoding/readState'
+import { OnChainData } from '../OnChainData'
+import { ByteWriter } from './ByteWriter'
+import { encodeForcedActions } from './encodeForcedActions'
+import { encodeOnChainData } from './encodeOnChainData'
+import { writeAssetConfigHashes } from './writeAssetConfigHashes'
+import { writeConditions } from './writeConditions'
+import { writeFundingEntries } from './writeFundingEntries'
+import { writeFundingIndices } from './writeFundingIndices'
+import { writeModifications } from './writeModifications'
+import { writeOraclePrices } from './writeOraclePrices'
+import { writePositionUpdate } from './writePositionUpdate'
+import { writeState } from './writeState'
 
 describe(encodeOnChainData.name, () => {
   const data: OnChainData = {
