+++ packages/encoding/src/test/data/spot-decoded-example.json
