+++ packages/encoding/src/test/data/spot-encoded-example.json
